Add support for TLS enabled endpoints
OpenStack supports endpoints that are protected by TLS.  URLs with
the https:// prefix can be found in the openrc file and also returned
from the keystone endpoint list.

VMTP, as a client, should be ready to send TLS enabled requests to
OpenStack and also verify the server side certificate that is sent
by the endpoint.

To validate a server side certificate, a client needs a copy of the
CA certificate representings the signing authority of the server
certificate.  For OpenStack, this CA certificat is pointed to by
the environment variable OS_CACERT.

The OpenStack client implementations accept the value in kwargs,
as either 'cacert' or 'ca_cert'.

Change-Id: I5b093d38ed51430d7c1ad09c65ec4961db818825