Reuse the router passed as config if available
It allows forcing the router and the external network because
the current process can easily detect a wrong external network
or router.

It detects if the router name set in config already exits. If yes,
it will be selected and the external network will be its gateway.
If the router is not connected to an external network, the process
will be unchanged. Then a new router will be created (same name).

It doesn't change the purpose of reuse_network_name (provider
networks).

Change-Id: I50069118fa7020900bea3dd579db18d087d3e47f
Signed-off-by: Cédric Ollivier <ollivier.cedric@gmail.com>